Convert entire project from Python to Node.js/TypeScript
It turns out it's very hard to create MCP Bundles that work across different
architectures and platforms when doing Python based MCP servers.

Since this project was started with the idea of testing Vibe Coding and had
large parts of the Python code implemented by Cursor, I decided to ask Cursor
to migrate everything to Node.js

This was quite a rollercoaster as Cursor broke a ton of things that had worked
in the Python version and also ended up checking many megabytes of dependencies
and build artifacts into the repo which I then had to remove again. But this
commit should represent a barely tested but at least not completely broken
first version of the Node.js implementation.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>